refactor(repositories): move repository related logic to the correct directory
