+++ modules/infusion/core/repositories/get-required-essentia.lua
